refactor: give the write endpoints their own exceptions module
The archive exception now sits in write/exceptions.py, alongside the versions package's
equivalent. Its serializers move to write/serializers.py so the exception can build its
response body without importing back from archive.py.
