+++ apps/api/v2/write/archive.py
@@ -6,69 +6,9 @@ to remove its webhook, and since this API cannot *create* a channel, a client th
 could never put it back. So a chatbot with a channel attached is refused, and a person detaches it.
 """
 
-from rest_framework import serializers, status
-from rest_framework.exceptions import APIException
-
+from apps.api.v2.write.exceptions import ChannelsAttached
 from apps.channels.models import ChannelPlatform, ExperimentChannel
 from apps.experiments.models import Experiment
-
-
-class AttachedChannelSerializer(serializers.Serializer):
-    """One channel standing in the way, named the way the web app names it.
-
-    Enough for a client to tell a person which channels to go and detach; there is nothing here to
-    address, because no endpoint takes a channel.
-    """
-
-    id = serializers.IntegerField()
-    # The same choice set the `chatbot_inspect` endpoint reports a channel's platform from, so a
-    # client parses one enum across both.
-    platform = serializers.ChoiceField(choices=ChannelPlatform.choices)
-    name = serializers.CharField()
-
-
-class ArchivedSerializer(serializers.Serializer):
-    """The archive response: that it happened, and what it destroyed on the way."""
-
-    archived = serializers.BooleanField(help_text="Always true; the failure cases are status codes.")
-    cancelled_scheduled_messages = serializers.IntegerField(
-        help_text=(
-            "How many of the chatbot's scheduled messages were deleted, finished and "
-            "already-cancelled ones included. They cannot be recovered."
-        )
-    )
-
-
-class ChannelsAttachedSerializer(serializers.Serializer):
-    """The 409: why the archive was refused, and what a person has to do about it."""
-
-    detail = serializers.CharField()
-    channels = AttachedChannelSerializer(many=True)
-
-
-class ChannelsAttached(APIException):
-    """The chatbot is still live on channels only a person can detach.
-
-    409 rather than 403: the caller may archive this chatbot, and will be able to, once the
-    channels are off it. Retrying without that changing will not help.
-    """
-
-    status_code = status.HTTP_409_CONFLICT
-
-    def __init__(self, channels: list[ExperimentChannel]) -> None:
-        # Assigned rather than handed to ``super().__init__``, which runs a structured detail
-        # through ``_get_error_details`` and turns every leaf into an ``ErrorDetail`` -- a ``str``
-        # subclass, so each channel's integer id would render as a string. Setting ``detail`` is
-        # all that ``__init__`` does with it, and the exception handler serves it as the body.
-        self.detail = {
-            "detail": (
-                f"This chatbot is still attached to {len(channels)} channel(s). Detaching one "
-                "removes its webhook at the messaging provider, so a person has to do it in the "
-                "web app -- this API has no channel endpoints. Archive again once they are "
-                "detached."
-            ),
-            "channels": AttachedChannelSerializer(channels, many=True).data,
-        }
 
 
 def archive_chatbot(chatbot: Experiment) -> dict:
